Change GrandTotal and TotalPrice to decimal for precision
Changed the type of the GrandTotal property in the CartIndexViewModel
class from double to decimal to improve precision in financial
calculations. Similarly, changed the type of the TotalPrice property
in the CartItemViewModel class from double to decimal for the same
reason.

+++ ST10361554-CLDV6212-POE-Part-3-WebApp/ViewModels/CartIndexViewModel.cs
@@ -6,6 +6,6 @@
         public List<CartItemViewModel>? CartItems { get; set; }
 
         // Total price of all items in the cart
-        public double GrandTotal { get; set; }
+        public decimal GrandTotal { get; set; }
     }
 }
